Add exceptions for metadataextractor, fix merge
