+++ converters/extractor/metadataextractor.py
@@ -13,6 +13,8 @@ import platform
 from helpers.utils import CommandExecutor, Formatter
 from converters.extractor.extractor import Extractor
 import os
+from exceptions.cmd_execute_exception import CmdExecutionError
+from exceptions.metadata_extract_exception import MetadataExtractationError
 
 
 class MetadataExtractor(Extractor):
@@ -27,6 +29,10 @@ class MetadataExtractor(Extractor):
         # Build the absolute path of the exiftool executable:
         exif_tool_path = cmd.get_execfile_path( "converters/extractor/bin/exifTool/" + exif_tool_distribution)
         exif_tool_command = f"{exif_tool_path} {self.file_path}"  # Create the command to run exiftool
-        result_text = cmd.run_command(exif_tool_command)
+        try:
+            result_text = cmd.run_command(exif_tool_command)
+        except CmdExecutionError as e:
+            raise MetadataExtractationError(f"Exiftool command execution failed: {e.get_message()}", e.get_status_code())
         formatter = Formatter()
         return formatter.key_value_string_to_dict(result_text, ":", "\n")
+
