+++ exceptions/convert_exception.py
@@ -1,13 +1,9 @@
-class ConvertError(Exception):
+from exceptions.base_exception import Error
+
+class ConvertError(Error):
     # Exception for convertion errors
     def __init__(self, message, status_code):
-        super().__init__(message)
-        self.status_code = status_code
-        self.message = message
+        super().__init__(message, status_code)
+        self.message = "Convertion error: " + message
 
-    def get_message(self):
-        return self.message
-    
-    def get_status_code(self):
-        return self.status_code
