+++ helpers/utils.py
@@ -1,6 +1,7 @@
 import subprocess
 import os
 import errno
+from exceptions.cmd_execute_exception import CmdExecutionError
 
 
 class CommandExecutor:
@@ -15,8 +16,8 @@ class CommandExecutor:
         try:
             return subprocess.run(command, shell=True, check=True, 
                                   encoding='utf-8', capture_output=True).stdout
-        except subprocess.CalledProcessError:
-            raise
+        except subprocess.CalledProcessError as e:
+            raise CmdExecutionError(f"Command {command} execution failed: {e.stderr.decode()}", 500)
         
 class Formatter:
     # Formats a string containing key:value pairs into a dictionary
